Use TbStringPool for all entity fields
