+++ common/data/src/main/java/org/thingsboard/server/common/data/edqs/fields/FieldsUtil.java
@@ -289,7 +289,7 @@ public class FieldsUtil {
     }
 
     public static String getText(JsonNode node) {
-        return node != null ? node.toString() : "";
+        return node != null && !node.isNull() ? node.toString() : "";
     }
 
     private static UUID getCustomerId(CustomerId customerId) {
