+++ common/util/src/main/java/org/thingsboard/common/util/TbStringPool.java
@@ -13,23 +13,21 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.thingsboard.server.edqs.util;
+package org.thingsboard.common.util;
 
-import com.google.common.hash.Hashing;
 import org.springframework.util.ConcurrentReferenceHashMap;
 
 import java.util.concurrent.ConcurrentMap;
 
-public class TbBytePool {
+public class TbStringPool {
 
-    private static final ConcurrentMap<String, byte[]> pool = new ConcurrentReferenceHashMap<>();
+    private static final ConcurrentMap<String, String> pool = new ConcurrentReferenceHashMap<>();
 
-    public static byte[] intern(byte[] data) {
+    public static String intern(String data) {
         if (data == null) {
             return null;
         }
-        var checksum = Hashing.sha512().hashBytes(data).toString();
-        return pool.computeIfAbsent(checksum, c -> data);
+        return pool.computeIfAbsent(data, str -> str);
     }
 
     public static int size(){
